Fix a wrong remove of ShiroUser when delete it into the manage user page + improve : cannot delete yourself (manage user) in order to prevent a permission error page.
git-svn-id: https://svn.asyd.net/svn/ghrm/trunk@34 91db7de8-7983-4966-9c1d-ef8a34a1989e

+++ grails-app/controllers/com/opencsi/ghrm/controller/UserController.groovy
@@ -1,5 +1,6 @@
 package com.opencsi.ghrm.controller
 
+import com.opencsi.ghrm.services.UserService
 import com.opencsi.ghrm.domain.User
 import com.opencsi.security.ShiroUser
 import com.opencsi.security.ShiroRole
@@ -98,15 +99,23 @@ class UserController {
 
 
     def delete = {
-        def userShiro = ShiroUser.get(params.id)
         def user = User.get(params.id)
-        if ( (user)&&(userShiro) )
+        def currentUser = new UserService()
+        if (currentUser.getAuthenticatedUserName() == user.toString())
         {
-            user.delete()
-            userShiro.delete()
-            flash.message = "User deleted."
-        }else
-            flash.message = "User deosn't exist."
+            flash.message = "Cannot delete your self!"
+        }
+        else
+        {
+            def userShiro = ShiroUser.findByUsername(user.toString())
+            if ( (user)&&(userShiro) )
+            {
+                user.delete()
+                userShiro.delete()
+                flash.message = "User deleted."
+            }else
+                flash.message = "User doesn't exist."
+        }
         redirect(action: "list")
     }
 }
